content/base/test/test_bug544642.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bug 544642</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a href="https://bugzilla.mozilla.org/show_bug.cgi?id=544642"
   target="_blank" >Mozilla Bug 544642</a>
<p id="display"></p>
<iframe id=iframe></iframe>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.8">
SimpleTest.waitForExplicitFinish();
var gen = runTest();

addLoadEvent(function() { gen.next(); });

function runTest() {
  var iframe = $('iframe');
  iframe.onerror = function() { gen.send("error"); };
  iframe.onload = function() { gen.send("load"); };

  iframe.src = "data:text/plain,hello";
  is((yield), "load", "plaintext data");

  iframe.src = "file://foo/bar";
  is((yield), "error", "file");

  // We should do this test too, however it brings up a modal dialog which
  // we can't dismiss.
  //iframe.src = "http:////";
  //is((yield), "error", "invalid http");

  SimpleTest.finish();
  
  yield undefined;
}

</script>
</pre>
</body>
</html>
